auth

+++ Frontend/src/app/modules/login/login.component.ts
@@ -1,7 +1,8 @@
 import { Component } from '@angular/core';
 import { Router } from '@angular/router';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
-import { AuthService } from 'src/app/services/auth.service';
+import { AuthService } from 'src/app/auth/auth.service';
+import Swal from 'sweetalert2';
 
 
 @Component({
@@ -28,10 +29,31 @@ export class LoginComponent {
 
   getOtp() {
     let value = this.emailLogin.value;
-    this.auth.beforeOTP(value).subscribe(data => {
-      console.log(data);
-      localStorage.setItem('email', value.email)
-      this.showDiv = !this.showDiv;
+    this.auth.beforeOTP(value).subscribe((res:any) => {
+      console.log('hi')
+         if (res.status) {
+      Swal.fire({
+        icon: 'success',
+        title: ' Sent',
+        showConfirmButton: false,
+        timer: 1500
+      }).then(() => {
+        this.emailLogin.reset();
+        localStorage.setItem('email', value.email)
+        this.showDiv = !this.showDiv;
+      })
+    }
+    else {
+      Swal.fire({
+        icon: 'error',
+        title: 'res.error',
+        showConfirmButton: false,
+        timer: 1500
+      }).then(() => {
+        this.emailLogin.reset()
+      })
+    }
+
 
     })
   }
@@ -40,12 +62,29 @@ export class LoginComponent {
     let otp = this.otpLogin.value.otp;
     let email = localStorage.getItem('email');
     let value = { email, otp }
-    this.auth.afterOTP(value).subscribe(data => {
-      console.log(data)
-      this.router.navigate(['/user']);
+    this.auth.afterOTP(value).subscribe((res:any) => {
+      if (res.status) {
+        Swal.fire({
+          icon: 'success',
+          title: ' Sent',
+          showConfirmButton: false,
+          timer: 1500
+        }).then(() => {
+          this.otpLogin.reset();
+          this.router.navigate(['/user']);        })
+      }
+      else {
+        Swal.fire({
+          icon: 'error',
+          title: 'res.error',
+          showConfirmButton: false,
+          timer: 1500
+        }).then(() => {
+          this.otpLogin.reset()
+        })
+      }     
     })
 
   }
 
-  onSubmit() { }
 }
